requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
python-dateutil>=2.8.0
tqdm>=4.64.0
faker>=18.0.0

# GPU acceleration (optional but recommended)
# Install appropriate version for your CUDA installation:
# For CUDA 11.x: cupy-cuda11x>=11.0.0
# For CUDA 12.x: 
cupy-cuda12x>=12.0.0
# For ROCm (AMD): cupy-rocm-5-0>=11.0.0

# Development and testing (optional)
pytest>=7.0.0
jupyter>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Performance profiling (optional)
memory-profiler>=0.60.0
line-profiler>=4.0.0